requirements.txt
numpy==2.2.6
opencv_python_headless==4.12.0.88
OpenEXR==3.4.4
pandas==2.3.3
Pillow==12.0.0
seaborn==0.13.2
